requirements.txt
numpy==1.24.3
Flask==2.2.2
Werkzeug==2.2.2
python-dotenv==1.0.1